+++ NEWS.md
@@ -6,7 +6,8 @@
 - Add active bindings for many of the waiter components,
 see `?Waiter`.
 - Deprecate functions with underscores in favour of camel case
-(only for that which are to be used in the browser).
+(only for that which are to be used in the shiny UI).
+- Deprecate the `logo` argument.
 
 # waiter 0.2.2